+++ lib/contexts/praise/events/events.dart
@@ -1,5 +1,5 @@
 import '../../../core/core.dart';
 
-class PraiseSentEvent extends ApplicationEvent<void> {
+class PraiseSentEvent extends ApplicationEvent<bool> {
   const PraiseSentEvent(super.data);
 }