+++ lib/main.dart
@@ -1,8 +1,8 @@
-import 'package:embrace/embrace.dart';
 import 'package:flutter/foundation.dart';
 import 'package:flutter/material.dart';
 import 'package:hive_flutter/hive_flutter.dart';
 import 'package:onesignal_flutter/onesignal_flutter.dart';
+import 'package:sentry_flutter/sentry_flutter.dart';
 import 'core/core.dart';
 
 import 'app.dart';
@@ -19,5 +19,21 @@ void main() async {
   OneSignal.initialize(Env.oneSignalAppId);
   OneSignal.Notifications.requestPermission(true);
 
-  await Embrace.instance.start(() => runApp(const App()));
+  await SentryFlutter.init(
+    (options) {
+      options.dsn = Env.sentryDSN;
+      options.tracesSampleRate = 1.0;
+    },
+    appRunner: () {
+      FlutterError.onError = (details) {
+        Sentry.captureException(
+          details.exception,
+          stackTrace: details.stack,
+        );
+      };
+      runApp(
+        const App(),
+      );
+    },
+  );
 }