+++ changelog.md
@@ -1,3 +1,11 @@
+# 1.2.5
+## Added
+- Sentry
+- private praise
+
+## Removed
+- Embrace sdk
+
 # 1.2.4
 ## Added
 - new member selection on praise flow